fix(api-gateway): cover all no-prefill Claude models
